refactor(files): create files controller and router
