+++ apps/api/src/controllers/applicationController.ts
@@ -23,13 +23,11 @@ import { getDbInstance } from '@/db/index.js';
 import logger from '@/logger.js';
 import { type ApplicationListRequest } from '@/routes/types.js';
 import { applicationSvc } from '@/service/applicationService.js';
-import { filesSvc } from '@/service/fileService.ts';
-import { FilesService, type ApplicationRecord, type ApplicationService } from '@/service/types.js';
+import { type ApplicationRecord, type ApplicationService } from '@/service/types.js';
 
 import { failure, success, type AsyncResult } from '@/utils/results.js';
 import { aliasApplicationContentsRecord, aliasApplicationRecord } from '@/utils/routes.js';
 import { type UpdateEditApplicationRequest } from '@pcgl-daco/validation';
-import formidable from 'formidable';
 import { ApplicationStateEvents, ApplicationStateManager } from './stateManager.js';
 
 /**
@@ -207,75 +205,6 @@ export const rejectApplication = async ({ applicationId }: { applicationId: numb
 		return updatedResult;
 	} catch (error) {
 		const message = `Unable to reject application with id: ${applicationId}`;
-		logger.error(message);
-		logger.error(error);
-		return failure(message, error);
-	}
-};
-
-/**
- * Upload a file with an associated application
- * @param applicationId - The target applicationId to associate the uploaded file
- * @param file - File blob
- * @returns Success with file data / Failure with Error.
- */
-export const uploadEthicsFile = async ({ applicationId, file }: { applicationId: number; file: formidable.File }) => {
-	try {
-		const database = getDbInstance();
-		const filesService: FilesService = filesSvc(database);
-		const applicationRepo: ApplicationService = applicationSvc(database);
-
-		const applicationResult = await applicationRepo.getApplicationWithContents({ id: applicationId });
-
-		if (!applicationResult.success) {
-			return failure('Failed getting application information');
-		}
-
-		const application = applicationResult.data;
-		const { edit } = ApplicationStateEvents;
-
-		const applicationRecord: ApplicationRecord = { ...application, contents: null };
-		const canEditResult = new ApplicationStateManager(applicationRecord)._canPerformAction(edit);
-
-		if (!canEditResult) {
-			return failure('Invalid action, must be in a draft state', 'Invalid action');
-		}
-
-		const ethicsLetterId = application.contents?.ethics_letter;
-
-		const txResult = await database.transaction(async (tx) => {
-			let result;
-
-			if (ethicsLetterId && ethicsLetterId !== null) {
-				result = await filesService.updateFile({
-					fileId: ethicsLetterId,
-					file,
-					application,
-					type: 'ETHICS_LETTER',
-					transaction: tx,
-				});
-			} else {
-				result = await filesService.createFile({ file, application, type: 'ETHICS_LETTER' });
-			}
-
-			if (!result.success) {
-				return result;
-			}
-
-			const applicantResult = await applicationRepo.editApplication({
-				id: applicationId,
-				update: { ethics_letter: result.data.id },
-				transaction: tx,
-			});
-
-			if (!applicantResult.success) {
-				return result;
-			}
-			return result;
-		});
-		return txResult;
-	} catch (error) {
-		const message = `Unable to upload file to application with id: ${applicationId}`;
 		logger.error(message);
 		logger.error(error);
 		return failure(message, error);
