+++ apps/api/src/server.ts
@@ -31,6 +31,7 @@ import collaboratorsRouter from '@/routes/collaboratorsRouter.js';
 import { serverConfig } from './config/serverConfig.js';
 import logger from './logger.js';
 import authRouter from './routes/authRouter.js';
+import fileRouter from './routes/fileRouter.ts';
 import signatureRouter from './routes/signatureRouter.ts';
 import sessionMiddleware from './session/sessionMiddleware.js';
 
@@ -50,6 +51,7 @@ const startServer = async () => {
 	app.use('/applications', applicationRouter);
 	app.use('/signature', signatureRouter);
 	app.use('/auth', authRouter);
+	app.use('/file', fileRouter);
 
 	app.use(
 		`/${API_DOCS_PATH}`,
